feat(player): 实现主字幕区动态环境光贯通与毛玻璃质感
- 将底层动态环境光提升至布局根容器，贯穿视频与字幕区
- 主字幕区采用半透明磨砂质感与调和透明度，自然透射视频光晕
- 移除主字幕面板最小高度限制（minSize={0}），支持用户自由缩放高度
- 修复 React Hook 顺序调用问题
